requirements-dev.txt
# Development and Testing Dependencies

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
moto==4.2.9  # AWS mocking

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Type stubs
boto3-stubs[dynamodb,lambda,secretsmanager]==1.34.0
types-requests==2.31.0.10

# Security
bandit==1.7.5
safety==2.3.5

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
